style:增加 example 路由
