+++ example/lib/routes.dart
@@ -1,0 +1,8 @@
+class AppRoutes {
+  static const String dashboard = '/';
+  static const String batteryDetails = '/battery-details';
+  static const String lowBattery = '/low-battery';
+  static const String peerSelection = '/peer-selection';
+  static const String iotControls = '/iot-controls';
+  static const String eventLog = '/event-log';
+}
